requirements.txt
Flask>=2.3.0
opencv-python>=4.8.0
numpy>=1.21.0
Pillow>=9.0.0
scikit-image>=0.19.0
matplotlib>=3.5.0
werkzeug>=2.3.0
